move postProcessParameter to AbstractJavaCodegen
