+++ modules/openapi-generator/src/main/java/org/openapitools/codegen/languages/JavaResteasyServerCodegen.java
@@ -17,16 +17,12 @@
 
 package org.openapitools.codegen.languages;
 
-import io.swagger.v3.oas.models.Operation;
 import org.apache.commons.lang3.BooleanUtils;
-import org.apache.commons.lang3.StringUtils;
 import org.openapitools.codegen.*;
 import org.openapitools.codegen.languages.features.JbossFeature;
 import org.openapitools.codegen.meta.features.DocumentationFeature;
-import org.openapitools.codegen.utils.PostProcessUtils;
 
 import java.io.File;
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -170,11 +166,6 @@ public class JavaResteasyServerCodegen extends AbstractJavaJAXRSServerCodegen im
         }
 
         return objs;
-    }
-
-    @Override
-    public void postProcessParameter(CodegenParameter p) {
-        PostProcessUtils.postProcessParameter(p);
     }
 
     public void setGenerateJbossDeploymentDescriptor(boolean generateJbossDeploymentDescriptor) {
